requirements.txt
# FastAPI core
fastapi
uvicorn[standard]
python-multipart
pydantic
pyyaml
requests

# Resume parsing
pdfplumber
python-docx
PyMuPDF

# NLP + Embeddings
spacy
chromadb

# Torch (CPU build – stable and hackathon-safe)
torch==2.4.1+cpu
torchvision==0.19.1+cpu
torchaudio==2.4.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu


